ListOfStrings does not require all fields sequentially filled
Previously retrieval of values would stop on first missing value of
request. Thus the user cannot leave blank entries.

For HTML inputs it was ok, because all fields would be transfered on
request, when including the dropdown autocompleters, cleared fields
are not transmitted. Thus it is better to inspect the entire request.

CMK-7960

Change-Id: I16085ef23468f4fee9aef6d1c95a93c1b5936e23