Fix ListOfStrings disabling of dropdown after new element
Notably on the ConfigHostname valuespec, that when in list of strings
the last element would stop working with select2 after selection and
creation of a new element.

Problem was during copy, it includes not only element id, but select2
has its own id, thus when new element was cloned, it included that id
and select2 disabled the previous instance.

CMK-7960

Change-Id: Ib7c7c8f1ef191832fe18edd47b15ab71dceeb9d0