12835 FIX Service label conditions could lead to wrong rule matches
The matching of service rule sets (which are used for a lot of different
settings in Checkmk) could lead to wrong rule matching results in case service
labels were used.

This issue was only triggered in a specific condition: Multiple services on
different hosts on the same site with the same service description had to use
the exact same set of label keys (the text before the ":") with different label
values (the text after the ":").

In this case it could happen that the rule value of one service affected the
rule value of another service. For example, this could lead to services being
added to a service group which should not be added to that group.

Change-Id: I70cde22f100bb7bb7961f6ca29307efd9d423d15