Fix setup search and add regression test
The specific RequestContext of the setup search had no theme
object set. This lead to a bug that made the setup search fail
after a recent refactoring.

* This commit fixes the specific issue
* It also adds a test that performs a search on the real index of
  Checkmk with all plugins loaded. This test would have shown the
  issue.

Todo: Investigate the related type declarations in cmk.gui.globals. They
seem to be wrong. If they would be correct, they should've uncovered
this issue.

Change-Id: I4adb033a0498feb048428d480e233a1cf59b480c