12941 FIX REST API: convert known host attributes to well defined structure
Previously the outbound structure of host and folder attributes was not defined. The internal
structure of the values were directly passed through. This was confusing because the inbound
structure of these attributes was already defined, leading to 2 distinct structures of the same
thing.

This is no longer the case. This werk introduces documentation and validation/conversion of outbound
host and folder attributes.

Some attribute values may now be different due to this change. Please check your scripts.

Change-Id: I768c00e0bfd63a9f446c777e907efb011d107b47