12301 FIX Fix cache usage for clusters
The cache was not used properly for clusters.  This resulted
in performance problems such that, if an host was part of <i>n</i>
clusters, it would be contacted <i>n + 1</i> times.  Now,
we make sure that the host is contacted a single time and that
checking for the clusters use locally cached data.

CMK-7752

Change-Id: Id3d0513a67480d0e1924403acb5231612dddc1ea