12817 FIX mk_postgres: improved error message "Instance DATA not running"
A postgres instance will be detected as running
if a process is found that contains a postgress commandline
which contains the name of the env-file.
Otherwise the check "SQL Instance" will show the
following error message "Instance <INSTANCE_NAME> not running or postgres DATADIR name is not identical with instance name.".

Change-Id: Ia2ee0481798dba3c480cd5383f1d3cdadad0942d