Pass path to PackedConfigStore
This makes the signature of the store consistent with other stores and
caches, such that the FileCache or the SectionStore.  Indeed, what the
store actually *wants* is a path where to read and write instead of a
serial number that is changed immediately to a path.  Moreover, the
serial number is an implementation detail of the path that we should not
be passing around too much.

This patch adds a class method taking the serial in order to only incur
minimal changes to the callers.

Change-Id: Ia50a879b2a66797c8e3c2cc0f894ca4664fada43