12492 Fix add metric from service view to custom and forecast graphs
Problem arose because the add page popup had not updated its variable
names to match the new Valuespec that controls metric specifications.

Fix is to establish a relation between the used valuespec and the add
popup, so that if ever in the future that changes, the link is
noticeable.

We have abused the inheritance of the autocompleters, their value is a
string. Yet for the metric autocompleters, because we use also the
information of the host and service hints, and even the type of metric.
Value is a tuple. Mypy complains that overriding methods in such way is
wrong(I agree), but until we get a better solution, the value arguments
and method return types remain untyped. Fortunately/disapointingly mypy
is not capable to infer the types in that case and does not complain
about this malpractice.

SUP-6705

Change-Id: Ic397bf0c925df3adc21aa2a675071233839b5437