Forecast Graphs: Fix upper bound when trend goes negative
This changes the representation. Now the lower bound does not exist on
the plot, the lower bound is just for the reference. Then the confidence
interval can be a floating stack-area, that means truly an interval.

Original bug, because the curve areas flip the lower,upper edge to keep
drawing order, and the staking arises from the upper bound, when trend
goes negative, then the upper bound stops following the trend and only
raises above the X axis.

Change-Id: Ibcc38c7390a62cfdb7d9d9c73b25f5131c1a1c8a
(cherry picked from commit 81b218f5f04ffcab285fb38821049da85ea6f899)