oracle_jobs: ValueError for setups with CDB and nonCDB on same host
If a host has multiple Databases with CDB and nonCDB Databases
with enabled oracle_jobs sections, all services for scheduler jobs
will crash due to wrong comparison for a nonCDB services against a CDB
agent line.

This bug has been introduced with the support for Multitenant Option
in 2018 and is fixed now.