HTML: Introduce new escaping methods
These escaping helpers basically perform the two different types
of escaping we use (strict and permissive) and don't return a str
object, but wrap the result in a HTML object. This makes the rest
of the code know, that this has already been escaped.

This is a preparation for further HTML / str related cleanups and
tightening of the HTML APIs.

Change-Id: I2fe35ca365c058d81e2c1c559572cfd2c995f2c6