HTML: Add strict escaping when joining with str now
Operations like HTML("a") + "b" are still allowed to keep call sites
clean, but the given str objects will now be HTML escaped.

In case you experience issues with escaped HTML after this change in the
GUI, you will likely have to clean up the call sites to provide the HTML
code wrapped in a HTML object instead of just handing over a str.

Another way would be to tighten the API of HTML and prevent the joins
with str, but we would have to update a lot of call sites and this would
also make the call sites more verbose.

Change-Id: Ic7307206e9c0a81fdb3fbf3151afa63d660b8fc4