Add type hints to legacy perfometers
This change uncovers a number of bugs from the type checking point of
view which are also fixed along the way.

Some of them seem like real bugs, but some of them may not appear in
reality because the data produced by the related checks prevents that.
However, it's a good idea to clarify all this.

Detail: Exceptions in perfometers are normally silently suppressed. It
is likely that these errors were not uncovered before for this reason.

Change-Id: I4a5128465748280b328cb4e48bd1f223d1271a1f