oracle_recovery_status: offline datafiles are ignorred from 12.1+
This is a critical bugfix for all environments with physical standby
databases, because informations from offline datafiles were removed from
the plugin output. Instead of CRIT state the informations were ignored
and the check is stale, when all daafiles are offline.
The has been fixed.

Important Note:

Please update the mk_oracle plugin for this fix!