12303 FIX Fix cache miss on discovery
Let cache max age depend on the mode

This changes the type of `max_age` from `int` to its own type in
the fetcher caches and the other changes use mypy and are mechanical.

This fixes a problem with the cache invalidation of the
discovery data.  This could have two consequences:
a performance problem where the hosts would be unnecessarily
queries; or the opposite where the data would not
be refreshed.

CMK-8136

Change-Id: Ifa046669f87c5f8d54e52f00c422693f8766339b