12945 REST API: option to show folders recursively and return type
The folder list endpoint now supports a "recursive" option, which
will result in the folder and all its subfolders to be in the resulting
collection.

Additionally, the endpoint will now always return the folder object and
never only the link. This makes it easier for script writers to parse
the result.

Change-Id: I655de590d7d06835e37a32b80282a99b14ce3530