12948 FIX REST API ETag handling
In 2.0.0p6, the viewing of a host through the REST API might fail
due to missing "modification time" information on the host. This led
to a HTTP 500 error. Also, the calculation of the ETag was sometimes
based on stale cached values, resulting in subsequent HTTP 412
"precondition failed" errors. This werk fixes both of these bug.

Change-Id: I10e2926dc2721af198a31fbef5d65ca2d56971fb