check_temperature: unique_name and value_store are now mandatory
The arguments unique_name and value_store are used to determine the
place where trend data is stored if the user has enabled trend
computation.

Because the arguments were optional before, it was easy to forget one
or both of the arguments when a trend computation was configurable
(e.g. if a plugin uses the ruleset "temperature"). In this case the
plugins seems to work, but crashes as soon as the user activates the
trend computation. This was e.g. the case for the migrated plugin
entity_sensors.

brocade_optical on the other hand did not crash since its ruleset does
not contain a trend option. Nonetheless it specified a unique_name
which was not strictly necessary.

Now unique_name and value_store are both required. This minimizes the
risk to misuse the check_temperature function. Although plugins like
brocade_optical do not need the trend computation right now it will be
easier to add this option later on since check_temperature will
correcly handle this case as soon as the ruleset is extended.

Change-Id: I5e8e92b355c14ac10c2cc5d41993a668136141bd