Bump urllib3 1.25.10 to 1.26.5
There are changes regarding TLSv1 and TLSv1.1 to urllib3 which may be
incompatible. We'll have to check this before porting it to 2.0.0.

These updates were done to get upstream security updates.

To make the urllib3 update possible, we have to update the following:

requests 2.25.1
certifi 2021.5.30
chardet 4.0.0
idna 2.10
botocore 1.20.102
boto3 1.17.102
s3transfer 0.4.2
jmespath 0.10.0
six 1.16.0

Change-Id: I40183a077c87756dfd6b10b0c7bba515d85e77af