12836 Disable outdated extension packages during site updates
Since 1.6 there is the information "valid until version" available in MKPs.
Since 2.0.0 Checkmk is now automatically disabling the packages during the site
update (omd update).

The mechanism tests all installed packages, checks the "valid until version"
information of each package and once it is set, compare the version with the
new Checkmk version you just updated to.

In case it is considered outdated, the package is moved to the disabled
packages. From this state you can either remove the package or reinstall it.

Change-Id: I5324fd23d98c44f2b0b2869e6bacc1703ded2adc