12983 FIX Setup search: Automatically recognize changes from extension packages
When installing, uninstalling, enabling or disabling an MKP, Checkmk
will now automatically re-build the index used in the search function
in the setup menu (in the background). Once this is complete, newly
added rulesets will be available in the setup search and removed
rulesets will no longer be displayed.

CMK-8185

Change-Id: I4130edc86cc4b886f2ae86fdae0c62b3ccbbd3f1