12903 FIX Management boards: cache DNS lookup
If provided with a trueish value for `configured_ip_address`
lookup_ip_address will not do any lookup, and return that
value.

As a result, this call would either return the management_address
unchanged (if not None) or resolve the *host* name.
While this always lead to the right result, it prevented the
resolution of the management name from being cached.

We now resolve the management address, if it is not already
an IP address.

Change-Id: Idd18f80e5c33d501385fb5ddbb71e68ae5c17c32