12918 FIX ups_capacity: add missing perfometer
The perfometer that was present in version 1.6 was missing in version 2.0.
The perfometer is now available again and changed to display the reamaining
time instead of the percentage. This seems to be more useful as the remaining
percentage may vary a lot. 100% may e.g. represent 1 or 100 minutes. This way
it is now possible to order the services by the remaining time.
Furthermore, the colors of the metrics are adapted to improve readability.

Change-Id: I72345d6d3d6dfb50ec5951c394ccec5624753c08