12838 FIX Fix error in reports when livestatus issues occur during report creation
When a report is being created and the reporting engine recognizes livestatus
connection issues to a site, error messages are added to the beginning of the
report to inform the user about the connection issue which may cause missing
data.

However, this mechanism lead to an exception in some cases with an error like
this: "Exception: KeyError ('site')".

CMK-8217

Change-Id: I79e98563eac41e78876edb5ca0fa331fe3711ef8