12919 FIX lsi_array, lsi_disk: fix crash of RAID array and RAID disk services
Due to a compatibility issue of the plugins with the new check plugin API
the services of the lsi_array plugin and the discovery service crashed after
an update to the 2.0. On the service discovery page all services of the
plugins were shown as vanished.

The services are now discovered correctly again. You have to rediscover the
services on the affected hosts.

SUP-6929

Change-Id: If2b8f4fd56e1cd239da4821b4d52e40dea9ccf1f