12839 FIX Background job details: Do not flicker while job is running
While a background job is running, the background job detail page shows
the current state of the background job and reloaded the information
once a second. During this update, the page flickered in some browsers
and also depending on the time needed to render the page.

The update mechanism of this page has now been reworked to not produce
these visual glitches.

Change-Id: I4101ab210393d092ce174c992951a7cac338353e