Fixed wrong pointer type.
There was a confusion between a std::unique_ptr<Foo> and a Foo*.
This worked just by accident: In libstdc++, a std::unique_ptr is
implemented as a tuple of the raw pointer and the deleter.
Furthermore, the std::tuple implementation effectively removes
empty elements where possible.

The net effect for all of this is: A std::unique_ptr<Foo> and a
Foo* have an identical representation in memory for libstdc++,
but this is just an implementation detail not guaranteed by the
standard.

Change-Id: I667289df2dde5d469cee8342bb7fda3491cb1d46