GUI config: Unglobalize parts of config loading
This is the first step on the way to unglobalize the config
handling of the GUI.

With this step applying of the default config, reading the config
files and applying their values is moved to a local `raw_config`
dictionary.

At the end, this `raw_config` dictionary is still applied to the globals
for compatibility. Once the rest of the code is refactored, this will
be dropped.

Change-Id: I553e5d74099ec78dac9c034730777133f0d96191