Drop the mandatory filter separation
This reduces complexity. The warning comes on each dashlet. There is no
ignore process on the filter list, which all other pages don't use.

The filter lists gets autocompleted because the context is forced to
include at least information of mandatory filters.

The VisualFilterList does not have the ignore feature anymore. That is
not only relevant for the FilterMenu but for all forms. The second use
was on filter forms separating single and multi context, yet in those
case the info_list did the job of separating repetition and forms still
look good. (Unless I'm missing something) I assume it works because in
those cases the filter name matches the single_info key, and if they are
already separated for the form, then there is no need for the ignore.

CMK-8232

Change-Id: I75cbea16520194e874ce1b5c83819910db52dae9