13022 Fix adding temperature graph to graph collections
Before this werk it was not possible to add a "Temperature Zone N" graph
to a graph collection. There was no error message visible, it was simply
not added to the collection.

The function used to receive the graph definition was only expecting
ascii characters, although the graph definition contained a degree
symbol.

Now a unicode compatible function is used to receive the graph
definition.

Change-Id: If77f940e609b12cbf301218b68ed005af9bce3bd