Dashboard plugins: Do not rely on the config
Even if the code tried to involve the user config, this did not
really work. We were always getting either the default config
or the config which was loaded by the previous apache request.

To cleanup this missleading structure, we now make clear that
the default permissions base on the default permissions of
the default roles.

Change-Id: I745fe3aaf58a4cc9f6ed8191e44c5c7d460a9f4e