12986 FIX <tt>apt</tt>: Handle Security Updates Line
Newer versions of the Linux command <tt>apt</tt> display something like
"10 standard security updates" or "3 esm-infra security updates" if security
updates are available. Up to now, the corresponding check plugin <tt>apt</tt>
was not able to handle this line. If such a line was present, the service
"APT Updates" was not discovered or went stale if already discovered.

This has been fixed.

Closes #384

Change-Id: Icb05d4e3cc8a7af886d08241b3d530acff4384f4