12927 SAP HANA: sap_hana_fileinfo and sap_hana_fileinfo_groups checks
Previously fileinfo and fileinfo_groups were used for monitoring
SAP HANA files and file groups.
Now new sap_hana_fileinfo and sap_hana_fileinfo_groups checks are used
for the same purpose. The difference is that with new checks services
go to STALE when there is no data (connection to the database is lost).

This werk is marked as incompatible because historical data from the
file monitoring will be lost when new checks are used.

Change-Id: I5e5fbd5453be0af56979cc98cd37ed1f780e1c32