12951 FIX response format of all host_config/folder_config REST API endpoints
The response format of some host_config and folder_config endpoints
did not fully conform to the OpenAPI spec. This has now been fixed, but
this also means that the following endpoints

   create folder
   hosts of folder
   update folder
   bulk update folders
   move a folder

   bulk create hosts
   list hosts

will have a different format in SOME attributes on the attributes key
within the "extensions" key. The now documented format in the OpenAPI
documentation is the correct one now.

Change-Id: I340943702cca54ccbf401e3f6fc9f8ed568303ac