Cleanup ntop test setup
Previously there was a local application_and_request_context created
and some manual monkeypatch.undo performed. This was necessary because
the setup steps were in the wrong order.

The single setup steps are now separated and executed in the correct
order.

The users are now also really created, just like we do it for the other
test users we need during testing. To enable this, the setup logic
for user accounts has been moved to the testlib.

Change-Id: I5bf2a27e1cd260ab5762a5500271beec821a3a9b