FIX linked_views_url_to have active filters
fixed linked urls and multi context report export

This is a bad solution, too much code repetition to pass active filters.
It should maybe be part of the makeuri.

The translation of filters also shows how weak the single_info is on the
url_vars. There is a difference on the filter ident and the
filter_html_var. Both need to be translated. There is also a KeyError
exception, which I believe under the correct translations should not be
required. It should not be possible from out side to create linked views
that would still be missing some context.

Change-Id: I1d475b87cdcbd4aae33a4a43e65d6b2bdfa9fb18