Introduce config.Config object
Previously the module namespace `cmk.gui.config` was used for holding the
GUI configuration during runtime. This is obviously error prone, bad for
testability and not thread safe.

This change introduces the `config.Config` object which will hold the
GUI configuration during runtime in the future.

For each request context we start with an empty `Config` object for now.
It holds only the factory settings. Once the config is loaded during
request processing, which is done by `config.initialize()`, at the
moment, the configuration will be accessible through
`cmk.gui.globals.config` which is a request local proxy object.

The `Config` classes attributes are defined by the builtin config
plugins from the different editions, which will be `CREConfig`,
`CEEConfig` and `CMEConfig`. Using the dataclasses allows us to have
code completion and type checked access to config variables, at least
for the builtin config variables.

For now the new config object is initialized and populated. Next steps:

* Add CME and CEE specific options
* Refactor all call sites to use the config object
* Drop the old structures which are currently needed for compatibility
* Rename "_" prefixed definitions in config plugins
* Improve config option type hints

Change-Id: Ic96e9b71f83208333d54f297aea45bde41bd24dc