13024 Restore SLA timeline proportions
The SLA timeline shows the state of a service by colored bars. Each
element's width should be in relation to the duration the service was in
this state.

Before this werk, the rect representing the state had a padding, so very
small time spans where falsely enlarged to bigger blocks.

This should now be fixed.

CMK-8188

Change-Id: I28ab023961e83e9be924a158e5b765b7d46e7038