13025 FIX make `cmk --snmptranslate` more robust
Before this werk it was possible that `cmk --snmptranslate` would crash
if the oids to translate where to long.

The oids where fed to snmptranslate as arguments and there is a limit
for the number of characters. If this limit was exceeded by too long
oids, `cmkd --snmptranslate` would stop without an error and needed to
be killed.

With this fix the oids are fed to snmptranslate via stdin. This
reduced the time to process a walk with 137379 lines from 3 minutes 35
seconds to 2.4 seconds.

As the general speed is now much better, the status output was removed.

Change-Id: I2b185bfdb0f2610dace097cc64aa40b40a194383