Update visuals context on major version update
VisualContext has a new strict form without single infos and everything
is a FilterHTTPVariables dict.

Although visuals(views/dashboards/reports) are updated on load (and
recursively for dashlets & report elements) by the GUI and consequently
after edit. It is helpful to also perform this update, on major version
update.
- Implement a context manager on config update script to save visuals.

Second update are Range Filters which now consistently use the "_until" suffix.
Filter renderer which does query out of the URL needs to query for the
optionally deprecated variable, because configuration might come out of
a Bookmark. General transformation from config files takes places with
the load proceduce as specified on previous comment.

Change-Id: I0ad5dd225a1be8311fd2c681af230f8503f8d96a