Single_info context transform
The context spec for single infos now includes a transform for in disk
to match the VisualContext type, which implies filters are separate. It
also needs a transform to extract from filter into single_spec
dictionary. (It works correctly but is so ugly that it can't be a good
option)

An extra transformation is provided when importing views previous to
this change(<2.1.0) to make sure context from disk is loaded correctly.
The transform_old_visuals acts on views, and recursively on
dashboard->daslet. Reports need an narrower step only for the context in
element.

Change-Id: Ib696f35601302d64f7550973ee89ca931dc7dc0d