13065 Absolute Levels on Filesystem Growth: Change Internal Representation From MB to Bytes
The absolute levels for filesystem growth trends used to be Integer
fields. This was changed to the Filesize field, to be more convenient
when inserting larger warning levels.

Custom plugins which use the `size_trend` functions might need some
minor adjustments since internally these levels are now handled as bytes
instead of MB. If no custom plugins are used, no actions are necessary.

Change-Id: Ia08b049e6ca6eb0be13717124dc6b6b12ffbc5e7