Align load_plugins fixture with real world
Previously only some tests were explicitly loading the GUI plugins.
Since the modules were loaded into the python and application global
structures, this is error prone. Subsequent tests which rely on the
plugins but not use the load_plugins fixture were working as long as the
previous test was executed before. But when the second test was executed
alone, this lead to a failing test.

To solve this situation the load_plugins fixture is now a session wide
fixture and automatically loaded for all GUI tests. This is less error
prone and also reflects the real world of the GUI code: The plugins are
loaded early in application initialization and available in the moment
the application code is executed.

Change-Id: I752b6bb6123a923facf9623421aee15837a4616d