13126 FIX postgres_conn_time: Don't crash on inconsistent section data
Previously, the postgres_conn_time check did crash while parsing inconsistent
section data.

Now, in the case of inparsable section data, but section data about the queried
database is available, a warning will be displayed in the resulting service.
Otherwise, if it is not possible to obtain information about the queried database,
it is assumed that no actual data is available, and the inconsistent section will
be dropped silently.

CMK-8342

Change-Id: I13b4d769e307986be5da2922f569ae4bd4a19188